update test for ubl validator
